fix!: Move backwards-compatibility hacks out of contextmenu.js
Move accessor property for ContextMenu.currentBlock from
core/contextmenu.js to core/main.js.
