+++ core/contextmenu.js
@@ -20,7 +20,6 @@ const Xml = goog.require('Blockly.Xml');
 const aria = goog.require('Blockly.utils.aria');
 const browserEvents = goog.require('Blockly.browserEvents');
 const clipboard = goog.require('Blockly.clipboard');
-const deprecation = goog.require('Blockly.utils.deprecation');
 const dom = goog.require('Blockly.utils.dom');
 const eventUtils = goog.require('Blockly.Events.utils');
 const userAgent = goog.require('Blockly.utils.userAgent');
@@ -70,32 +69,6 @@ const setCurrentBlock = function(block) {
   currentBlock = block;
 };
 exports.setCurrentBlock = setCurrentBlock;
-
-// Add JS accessors for backwards compatibility.
-Object.defineProperties(exports, {
-  /**
-   * Which block is the context menu attached to?
-   * @name Blockly.ContextMenu.currentBlock
-   * @type {Block}
-   * @deprecated Use Blockly.Tooltip.getCurrentBlock() /
-   *     .setCurrentBlock() instead.  (September 2021)
-   * @suppress {checkTypes}
-   */
-  currentBlock: {
-    get: function() {
-      deprecation.warn(
-          'Blockly.ContextMenu.currentBlock', 'September 2021',
-          'September 2022', 'Blockly.Tooltip.getCurrentBlock()');
-      return getCurrentBlock();
-    },
-    set: function(block) {
-      deprecation.warn(
-          'Blockly.ContextMenu.currentBlock', 'September 2021',
-          'September 2022', 'Blockly.Tooltip.setCurrentBlock(block)');
-      setCurrentBlock(block);
-    },
-  },
-});
 
 /**
  * Menu object.
