+++ core/events/events.js
@@ -15,7 +15,6 @@
  */
 goog.module('Blockly.Events');
 
-const deprecation = goog.require('Blockly.utils.deprecation');
 const eventUtils = goog.require('Blockly.Events.utils');
 const {Abstract: AbstractEvent} = goog.require('Blockly.Events.Abstract');
 const {BlockBase} = goog.require('Blockly.Events.BlockBase');
@@ -120,28 +119,3 @@ exports.isEnabled = eventUtils.isEnabled;
 exports.setGroup = eventUtils.setGroup;
 exports.setRecordUndo = eventUtils.setRecordUndo;
 exports.disableOrphans = eventUtils.disableOrphans;
-
-Object.defineProperties(exports, {
-  /**
-   * Sets whether the next event should be added to the undo stack.
-   * @name Blockly.Evenents.recordUndo
-   * @type {boolean}
-   * @deprecated Use Blockly.Events.getRecordUndo() and
-   *     .setRecordUndo().  (September 2021)
-   * @suppress {checkTypes}
-   */
-  recordUndo: {
-    get: function() {
-      deprecation.warn(
-          'Blockly.Events.recordUndo', 'September 2021', 'September 2022',
-          'Blockly.Events.getRecordUndo()');
-      return eventUtils.getRecordUndo();
-    },
-    set: function(record) {
-      deprecation.warn(
-          'Blockly.Events.recordUndo', 'September 2021', 'September 2022',
-          'Blockly.Events.setRecordUndo()');
-      eventUtils.setRecordUndo(record);
-    },
-  },
-});
